Revert "Merge branch 'master' of https://github.com/teckno/ivy"
This reverts commit 138bb765d228e89e2b477a6c9ddf0f3dc9b322d5, reversing
changes made to 966345fe69a40526586178047b508ecadc5a3bb3.

+++ ivy/functional/backends/torch/linear_algebra.py
@@ -136,7 +136,7 @@ def inner(
         x2 = x2.long()
     if ivy.exists(out):
         if out.dtype != x1.dtype:
-            return ivy.inplace_update(out, torch.inner(x1, x2).type(out.dtype))
+            return ivy.inplace_update(out, torch.inner(x1, x2).type(ret_dtype))
     return torch.inner(x1, x2, out=out).type(ret_dtype)
